requirements.txt
streamlit
pandas
scikit-learn
numpy
Pillow
seaborn
matplotlib
